change DGWStartTime to int64

+++ src/consensus/params.h
@@ -172,7 +172,7 @@ struct Params {
     int MidasStartHeight;
     int DGWStartHeight;
     int DGWDifficultyStartHeight;
-    int DGWStartTime;
+    int64_t DGWStartTime;
     int ATPStartHeight;
     int POSPOWStartHeight;
     /**
